fix: Change Schedule color field from Integer to BigInteger
Problem:
- Flutter Color.value is unsigned 32-bit (0 ~ 4294967295)
- PostgreSQL Integer is signed 32-bit (-2147483648 ~ 2147483647)
- Color values like 4294198070 exceeded Integer range
- Error: psycopg2.errors.NumericValueOutOfRange

Solution:
- Change color column from Integer to BigInteger in models.py
- Add migration script to alter existing database column
- BigInteger supports full unsigned 32-bit range for Flutter colors

Run migration:
cd backend && python migrate_fix_schedule_color.py

+++ backend/models.py
@@ -1,5 +1,5 @@
 # backend/models.py
-from sqlalchemy import Column, String, Text, DateTime, ForeignKey, Integer, Boolean, Float
+from sqlalchemy import Column, String, Text, DateTime, ForeignKey, Integer, Boolean, Float, BigInteger
 from sqlalchemy.orm import relationship
 from database import Base
 from datetime import datetime
@@ -111,7 +111,7 @@ class Schedule(Base):
     start_time = Column(String(5), nullable=True)  # HH:MM 형식
     end_time = Column(String(5), nullable=True)    # HH:MM 형식
     is_completed = Column(Boolean, default=False)
-    color = Column(Integer, nullable=True)  # Flutter Color.value (ARGB int)
+    color = Column(BigInteger, nullable=True)  # Flutter Color.value (ARGB unsigned 32-bit)
 
     user = relationship("User")
